Moves formMerkleTree and CreateHashFunction to merkleTreeService and updates module and profile
