+++ examples/bri-3/src/bri/merkleTree/services/merkleTree.service.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@nestjs/common';
+import MerkleTree from 'merkletreejs';
+import * as crypto from 'crypto';
+
+@Injectable()
+export class MerkleTreeService {
+  public formMerkleTree(leaves: string[], hashAlgName: string): MerkleTree {
+    const hashFn = this.createHashFunction(hashAlgName);
+    const hashedLeaves = leaves.map((x) => hashFn(x));
+    return new MerkleTree(hashedLeaves, hashFn);
+  }
+
+  public createHashFunction(hashAlgName: string): (data: any) => Buffer {
+    return (data: any): Buffer => {
+      return crypto.createHash(hashAlgName).update(data).digest();
+    };
+  }
+}
